web/requirements.txt
Flask
flask-restful
pymongo
bcrypt
numpy==1.18.5 # as tensorflow 2.3.1 doesn't support higher
tensorflow
requests
opencv-python
mtcnn
pandas
Pillow
